feat(admin): fold Performance and the procedural catalogues into two tabs
Two more groups of tabs that were one concern each.

PERFORMANCE. "Performance" and "Frontend Loads" split one question — why is it
slow — by which side is slow, which is precisely what the operator is trying to
find out. Picking the right tab required already knowing the answer. Now
Workers (cross-conversion cost) and Frontend (browser load/render) sit one
click apart, so comparing them is a comparison rather than a navigation.

PROCEDURAL ENGINE. Engines, System and Equipment are three per-scope catalogues
feeding the same consumer, the cellbuilder: which engine compiles a model,
which service systems it may route, which equipment it may place. One setup
task, done in one sitting, spread across a strip that said nothing about the
relationship. Ordered by what gets set up first.

Twelve top-level tabs become nine. With the audit fold earlier in this branch,
the strip has gone from fifteen to nine.

Three tabs now need a sub-tab strip, so it is one component instead of three
copies drifting apart. And the routing moved to a React-free `adminTabs`
module: the rules are the part worth testing, and importing AdminPanel in a
test pulls the whole viewer including a vite-only `?worker&inline` import that
no plain runner resolves. The test failing that way is what surfaced it.

Every retired hash still resolves — frontend_loads, equipment, system, engines,
alongside the audit ones — because they are in bookmarks, in history and in
in-app triggers. Nine tests pin exactly that, including the case that matters
most: an unknown SUB-tab falls back to its own tab's default rather than
silently landing on a different panel.

678 REST tests, 330 frontend tests, lint-check clean.
