+++ src/frontend/package.json
@@ -18,7 +18,7 @@
     "build:serve": "vite build --config vite.config.serve.ts --emptyOutDir",
     "build:embed": "vite build --config vite.config.embed.ts",
     "dev": "vite --force",
-    "test": "node --import tsx --test src/__tests__/services/wasmSupport.test.ts src/__tests__/services/boolOptionGating.test.ts src/__tests__/services/feaManifestPoll.test.ts src/__tests__/services/feaFieldBlob.test.ts src/__tests__/services/feaElemFieldBlob.test.ts src/__tests__/services/feaMeshEdges.test.ts src/__tests__/services/feaMeshElements.test.ts src/__tests__/services/feaBeamSolidsWarp.test.ts src/__tests__/services/oidcScopedToken.test.ts src/__tests__/utils/edgeShaderHelper.test.ts src/__tests__/utils/rangeIndex.test.ts src/__tests__/utils/customBatchedMeshHighlight.test.ts src/__tests__/utils/edgeOverlayTransform.test.ts src/__tests__/utils/sourceUpAxis.test.ts src/__tests__/utils/cellbuilderSnap.test.ts src/__tests__/utils/cellbuilderPorts.test.ts src/__tests__/utils/cellbuilderEquipmentPreviewBox.test.ts src/__tests__/utils/cellbuilderCompileGate.test.ts src/__tests__/utils/cellbuilderHistory.test.ts src/__tests__/utils/cellbuilderLoft.test.ts src/__tests__/utils/cellbuilderBlueprints.test.ts src/__tests__/utils/cellbuilderDetailingOptions.test.ts src/__tests__/utils/cellbuilderGroups.test.ts src/__tests__/utils/cellbuilderMetadata.test.ts src/__tests__/utils/cellbuilderSideBySide.test.ts src/__tests__/utils/typeIconClassify.test.ts src/__tests__/utils/portColor.test.ts src/__tests__/utils/treeGraph.test.ts src/__tests__/utils/treeNodeFacts.test.ts src/__tests__/utils/modelStats.test.ts src/__tests__/state/cellBuilderMetadata.store.test.ts src/__tests__/state/cellBuilderViewState.store.test.ts src/__tests__/components/detailingPanel.smoke.test.tsx src/__tests__/plugins/registry.test.ts src/__tests__/plugins/demoPlugin.test.tsx src/__tests__/plugins/uiShells.test.ts src/__tests__/plugins/viewerCoreFacade.test.ts src/__tests__/plugins/pluginStyles.test.ts src/__tests__/services/externalModelsBinding.test.ts src/__tests__/services/externalModelUpload.test.ts src/__tests__/services/externalModelClients.test.ts src/__tests__/services/externalModelDispatch.test.ts src/__tests__/state/auditFilter.store.test.ts"
+    "test": "node --import tsx --test src/__tests__/services/wasmSupport.test.ts src/__tests__/services/boolOptionGating.test.ts src/__tests__/services/feaManifestPoll.test.ts src/__tests__/services/feaFieldBlob.test.ts src/__tests__/services/feaElemFieldBlob.test.ts src/__tests__/services/feaMeshEdges.test.ts src/__tests__/services/feaMeshElements.test.ts src/__tests__/services/feaBeamSolidsWarp.test.ts src/__tests__/services/oidcScopedToken.test.ts src/__tests__/utils/edgeShaderHelper.test.ts src/__tests__/utils/rangeIndex.test.ts src/__tests__/utils/customBatchedMeshHighlight.test.ts src/__tests__/utils/edgeOverlayTransform.test.ts src/__tests__/utils/sourceUpAxis.test.ts src/__tests__/utils/cellbuilderSnap.test.ts src/__tests__/utils/cellbuilderPorts.test.ts src/__tests__/utils/cellbuilderEquipmentPreviewBox.test.ts src/__tests__/utils/cellbuilderCompileGate.test.ts src/__tests__/utils/cellbuilderHistory.test.ts src/__tests__/utils/cellbuilderLoft.test.ts src/__tests__/utils/cellbuilderBlueprints.test.ts src/__tests__/utils/cellbuilderDetailingOptions.test.ts src/__tests__/utils/cellbuilderGroups.test.ts src/__tests__/utils/cellbuilderMetadata.test.ts src/__tests__/utils/cellbuilderSideBySide.test.ts src/__tests__/utils/typeIconClassify.test.ts src/__tests__/utils/portColor.test.ts src/__tests__/utils/treeGraph.test.ts src/__tests__/utils/treeNodeFacts.test.ts src/__tests__/utils/modelStats.test.ts src/__tests__/state/cellBuilderMetadata.store.test.ts src/__tests__/state/cellBuilderViewState.store.test.ts src/__tests__/components/detailingPanel.smoke.test.tsx src/__tests__/plugins/registry.test.ts src/__tests__/plugins/demoPlugin.test.tsx src/__tests__/plugins/uiShells.test.ts src/__tests__/plugins/viewerCoreFacade.test.ts src/__tests__/plugins/pluginStyles.test.ts src/__tests__/services/externalModelsBinding.test.ts src/__tests__/services/externalModelUpload.test.ts src/__tests__/services/externalModelClients.test.ts src/__tests__/services/externalModelDispatch.test.ts src/__tests__/state/auditFilter.store.test.ts src/__tests__/components/adminTabRouting.test.ts"
   },
   "devDependencies": {
     "@tailwindcss/postcss": "^4.3.0",
